fix(audio): persist SoundCloud playback across navigation + now-playing artwork
The audio used to restart on every page change because each page renders its
own <Navigation> (which held the brightness-rack + the iframe), so it remounted.

- Move the SoundCloud iframe + widget + state into a SoundCloudAudioProvider
  mounted ABOVE the router in App.tsx, so playback persists across client-side
  navigation. brightness-rack reads state via useSoundCloudAudio() context.
- Playback view is now a 'now playing' card: track artwork (artwork_url, falling
  back to the profile avatar; upscaled to 200px) + title + 'pommeshdrms · genre'
  with discreet SoundCloud / Instagram links underneath + the mute toggle.
- Context/hook split into use-sound-cloud-audio.ts (react-refresh clean).
